Create Selection
You can now select up to one hex at a time by clicking on it. Clicking it again or selecting a different hex unselects it. The numbers 1-9 set the value of the selected hex to be that number, 0 sets the selected hex's value to 10, and backspace/delete clears the hex. Changed how state transitions are handled.