Better Controls
In game, you can now right click a hex to make it empty again in addition to the existing backspace and delete options. In the level editor, you can right click a hex to remove it from the level. Left clicking a removed hex in the level editor now brings it back instead of selecting it.